Conflicts:

	Client Applications/rcracki_mt/ChainWalkContext.cpp
	Client Applications/rcracki_mt/ChainWalkContext.h
	Client Applications/rcracki_mt/CrackEngine.cpp
	Client Applications/rcracki_mt/HashAlgorithm.cpp
	Client Applications/rcracki_mt/HashRoutine.cpp
	Client Applications/rcracki_mt/HashSet.cpp
	Client Applications/rcracki_mt/MemoryPool.cpp
	Client Applications/rcracki_mt/MemoryPool.h
	Client Applications/rcracki_mt/Public.cpp
	Client Applications/rcracki_mt/Public.h
	Client Applications/rcracki_mt/RTI2Reader.cpp
	Client Applications/rcracki_mt/RTI2Reader.h
	Client Applications/rcracki_mt/RainbowCrack.cpp
	Client Applications/rcracki_mt/lm2ntlm.cpp
	Client Applications/rcracki_mt/md4.cpp
	Client Applications/rcracki_mt/rcrackiThread.cpp
	Client Applications/rcracki_mt/sha1.cpp
	Client Applications/rti2rto/rti2rto.cpp
	Common/rt api/BaseRTReader.h
	Common/rt api/MemoryPool.cpp
	Common/rt api/RTI2Reader.cpp
	Common/rt api/RTIReader.cpp

merging with PB's changes
UINT4 -> uint32 changes
misc fixes for rti2
misc fixes for long 64/32 compat

+++ Client Applications/rcracki_mt/ChainWalkSet.cpp
@@ -67,7 +67,7 @@ string CChainWalkSet::CheckOrRotatePreCalcFile()
 		sprintf(sPreCalcFileName, "%s.%d", sPrecalcPathName.c_str(), preCalcPart);
 		string sReturnPreCalcPath(sPreCalcFileName);
 
-		unsigned int fileLen = 0;
+		long fileLen = 0;
 
 		FILE* file = fopen(sReturnPreCalcPath.c_str(), "ab");
 		if(file!=NULL)
